+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 UNRELEASED
 
+- Use round robin load balancing instead of 'pick first' for gitaly-ruby
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/700
 - Unvendor Repository#create implementation
   https://gitlab.com/gitlab-org/gitaly/merge_requests/713
 - Add gitaly-ruby installation debug log messages